Wip/rewrite

+++ lang/src_ref/line_col.rs
@@ -1,13 +1,13 @@
-// Copyright © 2024 The µcad authors <info@ucad.xyz>
+// Copyright © 2024-2025 The µcad authors <info@ucad.xyz>
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 /// Line and column within a source code file
 #[derive(Clone, Debug, Default)]
 pub struct LineCol {
     /// Line number (1..)
-    pub line: u32,
+    pub line: usize,
     /// Column number (1..)
-    pub col: u32,
+    pub col: usize,
 }
 
 impl std::fmt::Display for LineCol {
